Better support pull requests

+++ entrypoint.sh
@@ -3,11 +3,18 @@
 set -e
 
 # extract info
-# version=${GITHUB_REF}
-# version=${version#refs/tags/}
-# version=${version#v}
-version=dev
-
+if [[ "$GITHUB_REF" == refs/tags/* ]]; then
+  version=${GITHUB_REF#refs/tags/}
+  version=${version#v}
+  release=true
+elif [[ "$GITHUB_REF" == refs/heads/* ]]; then
+  version=${GITHUB_REF#refs/heads/}
+  release=false
+elif [[ "$GITHUB_REF" == refs/pull/* ]]; then
+  pull_number=$(jq --raw-output .pull_request.number "$GITHUB_EVENT_PATH")
+  version=pr-${pull_number}
+  release=false
+fi
 
 name=$(yq -r .final_name config/final.yml)
 if [ "${name}" = "null" ]; then
@@ -27,23 +34,30 @@ if [ ! -z "${INPUT_BUNDLE}" ] && [ "${INPUT_BUNDLE}" != "false" ]; then
   gem install bundler -v "${INPUT_BUNDLE}"
 fi
 
-
-# remove existing release if any
-if [ -f releases/${name}/${name}-${version}.yml ]; then
-  echo "removing pre-existing version ${version}"
-  yq -r -y "{ \"builds\": (.builds | with_entries(select(.value.version != \"${version}\"))), \"format-version\": .[\"format-version\"]}" < releases/${name}/index.yml > tmp
-  mv tmp releases/${name}/index.yml
-  rm -f releases/${name}/${name}-${version}.yml
-  git commit -a -m "reset release ${version}"
+if [ "${release}" == "true" ]
+  # remove existing release if any
+  if [ -f releases/${name}/${name}-${version}.yml ]; then
+    echo "removing pre-existing version ${version}"
+    yq -r -y "{ \"builds\": (.builds | with_entries(select(.value.version != \"${version}\"))), \"format-version\": .[\"format-version\"]}" < releases/${name}/index.yml > tmp
+    mv tmp releases/${name}/index.yml
+    rm -f releases/${name}/${name}-${version}.yml
+    git commit -a -m "reset release ${version}"
+  fi
 fi
 
 echo "creating bosh release: ${name}-${version}.tgz"
-bosh create-release --force --timestamp-version --tarball=${name}-${version}.tgz
+if [ "${release}" == "true" ]
+  bosh create-release --force --final --version=${version} --tarball=${name}-${version}.tgz
+else
+  bosh create-release --force --timestamp-version --tarball=${name}-${version}.tgz
+fi
 
-# echo "pushing changes to git repository"
-# git add releases/${name}/${name}-${version}.yml
-# git commit -a -m "cutting release ${version}"
-# git push ${remote_repo} HEAD:${INPUT_TARGET_BRANCH}
+if [ "${release}" == "true" ]
+  echo "pushing changes to git repository"
+  git add releases/${name}/${name}-${version}.yml
+  git commit -a -m "cutting release ${version}"
+  git push ${remote_repo} HEAD:${INPUT_TARGET_BRANCH}
+fi
 
 # make asset readble outside docker image
 chmod 644 ${name}-${version}.tgz
